working on swipe

+++ src/main/java/com/loop614/analyticstribble/tracker/transfer/VectorTransfer.java
@@ -1,14 +1,12 @@
 package com.loop614.analyticstribble.tracker.transfer;
 
-import java.math.BigDecimal;
-
 public class VectorTransfer {
 
-    public BigDecimal x;
-    public BigDecimal y;
-    public BigDecimal dt;
+    public Integer x;
+    public Integer y;
+    public Long dt;
 
-    public VectorTransfer(BigDecimal x, BigDecimal y, BigDecimal dt) {
+    public VectorTransfer(Integer x, Integer y, Long dt) {
         this.x = x;
         this.y = y;
         this.dt = dt;
